requirements.txt
pdfplumber>=0.10.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
openai>=1.0.0